requirements.txt
anyascii==0.3.2
asgiref==3.7.2
astroid==3.0.2
beautifulsoup4==4.11.2
black==23.12.1
boto3==1.34.34
botocore==1.34.34
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
Django==4.2.9
django-cors-headers==4.3.1
django-filter==23.5
django-modelcluster==6.1
django-permissionedforms==0.1
django-storages==1.14.2
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
djlint==1.34.1
draftjs-exporter==2.1.7
EditorConfig==0.12.3
environ==1.0
et-xmlfile==1.1.0
filelock==3.13.1
filetype==1.2.0
flake8==6.1.0
future==0.18.3
geocoder==1.38.1
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
identify==2.5.33
idna==3.6
isort==5.13.2
jmespath==1.0.1
jsbeautifier==1.14.11
json5==0.9.14
l18n==2021.3
Markdown==3.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openpyxl==3.1.2
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
pillow-heif==0.14.0
pip-review==1.3.0
platformdirs==4.1.0
pre-commit==3.5.0
psycopg==3.1.16
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.0.7
virtualenv==20.25.0
wagtail==5.2.2
webencodings==0.5.1
Willow==1.6.3
